FEAT: Added basic typing test functionality.
When a user types into the input form, the corresponding letter in the text is changed to either red or green depending on whether they are correct. If incorrect, their mistakes counter is incremented by one. Additionally, it counts the number of words the user has typed. Finally, we have imported a timer that counts down to 0.